[NER-868] Libutils' first commit is based on esp-idf-ansic1218 project.
In order to keep previous changes, the 'esp-idf-libtuils' repo was
based on 'esp-idf-ansic1218' as the serial and exceptions first
implementation was derived from there. Files not related with the
libutils were excluded and, from now on, the repository must
aggregate only files commonly used by other components of the system.